// WebClient/src/api/GeneratedApiClient.ts
export interface AgentResponse {
    speechFilename?: string | undefined;
    lipSyncFilename?: string | undefined;
}

export interface AuthModel {
    username?: string | undefined;
    password?: string | undefined;
}

export interface AvatarGender {
    type?: AvatarGenderTypeEnum;
}

export enum AvatarGenderTypeEnum {
    Male = "Male",
    Female = "Female",
}